Created a folder for all stat related components. I created components for each statistic that will be shown. More will be added in the future.
